+++ src/obu.rs
@@ -202,8 +202,8 @@ fn parse_seq_hdr(
     } else {
         timing_info_present = gb.get_bit() as u8;
         if timing_info_present != 0 {
-            num_units_in_tick = gb.get_bits(32) as u32;
-            time_scale = gb.get_bits(32) as u32;
+            num_units_in_tick = gb.get_bits(32);
+            time_scale = gb.get_bits(32);
             if strict_std_compliance && (num_units_in_tick == 0 || time_scale == 0) {
                 return Err(EINVAL);
             }
@@ -222,7 +222,7 @@ fn parse_seq_hdr(
             decoder_model_info_present = gb.get_bit() as u8;
             if decoder_model_info_present != 0 {
                 encoder_decoder_buffer_delay_length = gb.get_bits(5) as u8 + 1;
-                num_units_in_decoding_tick = gb.get_bits(32) as u32;
+                num_units_in_decoding_tick = gb.get_bits(32);
                 if strict_std_compliance && num_units_in_decoding_tick == 0 {
                     return Err(EINVAL);
                 }
@@ -267,9 +267,9 @@ fn parse_seq_hdr(
                 if op.decoder_model_param_present != 0 {
                     let opi = &mut operating_parameter_info[i as usize];
                     opi.decoder_buffer_delay =
-                        gb.get_bits(encoder_decoder_buffer_delay_length.into()) as u32;
+                        gb.get_bits(encoder_decoder_buffer_delay_length.into());
                     opi.encoder_buffer_delay =
-                        gb.get_bits(encoder_decoder_buffer_delay_length.into()) as u32;
+                        gb.get_bits(encoder_decoder_buffer_delay_length.into());
                     opi.low_delay_mode = gb.get_bit() as u8;
                 }
             }
@@ -816,7 +816,7 @@ fn parse_refidx(
             refidx[i as usize] = gb.get_bits(3) as i8;
         }
         if seqhdr.frame_id_numbers_present != 0 {
-            let delta_ref_frame_id = gb.get_bits(seqhdr.delta_frame_id_n_bits.into()) as u32 + 1;
+            let delta_ref_frame_id = gb.get_bits(seqhdr.delta_frame_id_n_bits.into()) + 1;
             let ref_frame_id = frame_id + (1 << seqhdr.frame_id_n_bits) - delta_ref_frame_id
                 & (1 << seqhdr.frame_id_n_bits) - 1;
             state.refs[refidx[i as usize] as usize]
@@ -1044,7 +1044,7 @@ fn parse_quant(
 
 fn parse_seg_data(gb: &mut GetBits) -> Rav1dSegmentationDataSet {
     let mut preskip = 0;
-    let mut last_active_segid = -1 as i8;
+    let mut last_active_segid = -1_i8;
     let d = array::from_fn(|i| {
         let i = i as i8;
         let delta_q;
@@ -1778,15 +1778,14 @@ fn parse_frame_hdr(
     if show_existing_frame != 0 {
         existing_frame_idx = gb.get_bits(3) as u8;
         if seqhdr.decoder_model_info_present != 0 && seqhdr.equal_picture_interval == 0 {
-            frame_presentation_delay =
-                gb.get_bits(seqhdr.frame_presentation_delay_length.into()) as u32;
+            frame_presentation_delay = gb.get_bits(seqhdr.frame_presentation_delay_length.into());
         } else {
             // Default initialization.
             frame_presentation_delay = Default::default();
         }
         let frame_id;
         if seqhdr.frame_id_numbers_present != 0 {
-            frame_id = gb.get_bits(seqhdr.frame_id_n_bits.into()) as u32;
+            frame_id = gb.get_bits(seqhdr.frame_id_n_bits.into());
             state.refs[existing_frame_idx as usize]
                 .p
                 .p
@@ -1824,8 +1823,7 @@ fn parse_frame_hdr(
     let showable_frame;
     if show_frame != 0 {
         if seqhdr.decoder_model_info_present != 0 && seqhdr.equal_picture_interval == 0 {
-            frame_presentation_delay =
-                gb.get_bits(seqhdr.frame_presentation_delay_length.into()) as u32;
+            frame_presentation_delay = gb.get_bits(seqhdr.frame_presentation_delay_length.into());
         }
         showable_frame = (frame_type != Rav1dFrameType::Key) as u8;
     } else {
@@ -1857,7 +1855,7 @@ fn parse_frame_hdr(
     }
 
     let frame_id = match seqhdr.frame_id_numbers_present {
-        0 => gb.get_bits(seqhdr.frame_id_n_bits.into()) as u32,
+        0 => gb.get_bits(seqhdr.frame_id_n_bits.into()),
         _ => 0,
     };
 
@@ -1894,7 +1892,7 @@ fn parse_frame_hdr(
                     let in_spatial_layer = seqop.idc >> spatial_id + 8 & 1;
                     if seqop.idc == 0 || in_temporal_layer != 0 && in_spatial_layer != 0 {
                         op.buffer_removal_time =
-                            gb.get_bits(seqhdr.buffer_removal_delay_length.into()) as u32;
+                            gb.get_bits(seqhdr.buffer_removal_delay_length.into());
                     }
                 }
             }
@@ -2278,11 +2276,11 @@ fn parse_obus(
                     let _ = mem::take(&mut state.content_light);
                     let _ = mem::take(&mut state.mastering_display);
                     for i in 0..8 {
-                        if state.refs[i as usize].p.p.frame_hdr.is_some() {
-                            let _ = mem::take(&mut state.refs[i as usize].p);
+                        if state.refs[i].p.p.frame_hdr.is_some() {
+                            let _ = mem::take(&mut state.refs[i].p);
                         }
-                        let _ = mem::take(&mut state.refs[i as usize].segmap);
-                        let _ = mem::take(&mut state.refs[i as usize].refmvs);
+                        let _ = mem::take(&mut state.refs[i].segmap);
+                        let _ = mem::take(&mut state.refs[i].refmvs);
                         let _ = mem::take(&mut state.cdf[i]);
                     }
                     state.frame_flags |= PictureFlags::NEW_SEQUENCE;
@@ -2413,7 +2411,7 @@ fn parse_obus(
                 Some(ObuMetaType::ItutT35) => {
                     let mut payload_size = gb.remaining_len();
                     // Don't take into account all the trailing bits for `payload_size`.
-                    while payload_size > 0 && gb[payload_size as usize - 1] == 0 {
+                    while payload_size > 0 && gb[payload_size - 1] == 0 {
                         payload_size -= 1; // trailing_zero_bit x 8
                     }
                     payload_size -= 1; // trailing_one_bit + trailing_zero_bit x 7
@@ -2431,7 +2429,7 @@ fn parse_obus(
                     } else {
                         let country_code = country_code as u8;
                         let country_code_extension_byte = country_code_extension_byte as u8;
-                        let payload = gb.get_bytes(payload_size as usize).into(); // TODO fallible allocation
+                        let payload = gb.get_bytes(payload_size).into(); // TODO fallible allocation
                         let itut_t35 = Rav1dITUTT35 {
                             country_code,
                             country_code_extension_byte,
